+++ src/conf_world/lib/SecondRoute.dart
@@ -1,0 +1,38 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_map/flutter_map.dart';
+import 'package:latlong/latlong.dart';
+import 'package:flutter/widgets.dart';
+
+class SecondRoute extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return new Scaffold(
+      appBar: new AppBar(title: new Text('Conf World')),
+      body: new FlutterMap(
+        options: new MapOptions(
+          center: new LatLng(51.5, -0.09),
+          zoom: 13.0,
+        ),
+        layers: [
+          new TileLayerOptions(
+              urlTemplate: "https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png",
+              subdomains: ['a', 'b', 'c']
+          ),
+          new MarkerLayerOptions(
+            markers: [
+              new Marker(
+                width: 80.0,
+                height: 80.0,
+                point: new LatLng(51.5, -0.09),
+                builder: (ctx) =>
+                new Container(
+                  child: new FlutterLogo(), // Aparece o marcador
+                ),
+              ),
+            ],
+          ),
+        ],
+      ),
+    );
+  }
+}
