+++ src/conf_world/test/widget_test.dart
@@ -10,6 +10,7 @@ import 'package:flutter_test/flutter_test.dart';
 
 import 'package:conf_world/main.dart';
 
+/*
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
@@ -28,3 +29,4 @@ void main() {
     expect(find.text('1'), findsOneWidget);
   });
 }
+ */
